remove computer abstract class, use player abstract class instead of interface, rename to runGame
