+++ src/main/java/tictactoe/cli/Main.java
@@ -3,8 +3,9 @@ package tictactoe.cli;
 public class Main {
 
     public static void main(String[] args) {
-        ConsoleMediator consoleMediator = new ConsoleMediator(new Console(System.in, System.out));
-        consoleMediator.startGame();
+        Console console = new Console(System.in, System.out);
+        ConsoleMediator consoleMediator = new ConsoleMediator(console);
+        consoleMediator.runGame();
     }
 
 }
