+++ src/main/java/tictactoe/core/Tile.java
@@ -1,6 +1,5 @@
 package tictactoe.core;
 
-import tictactoe.core.players.Player;
 import tictactoe.core.players.PlayerSymbol;
 
 import java.util.Optional;
