+++ src/main/java/tictactoe/core/players/HumanPlayer.java
@@ -3,10 +3,7 @@ package tictactoe.core.players;
 import tictactoe.core.Board;
 import tictactoe.core.Mediator;
 
-public class HumanPlayer implements Player {
-
-    private final PlayerSymbol playerSymbol;
-    private final Mediator mediator;
+public class HumanPlayer extends Player {
 
     public HumanPlayer(PlayerSymbol playerSymbol, Mediator mediator) {
         this.playerSymbol = playerSymbol;
@@ -16,11 +13,6 @@ public class HumanPlayer implements Player {
     @Override
     public void requestMove(Board board) {
         mediator.requestMoveFromUI(board, playerSymbol);
-    }
-
-    @Override
-    public PlayerSymbol getSymbol() {
-        return playerSymbol;
     }
 
 }
