+++ src/main/java/tictactoe/core/players/Players.java
@@ -29,4 +29,5 @@ public class Players {
             currentPlayer = this.player1;
         }
     }
+
 }
