+++ src/main/java/tictactoe/core/players/UnbeatablePlayer.java
@@ -4,34 +4,41 @@ import tictactoe.core.Board;
 import tictactoe.core.Mediator;
 import tictactoe.core.util.Time;
 
-public class UnbeatablePlayer extends Computer implements Player {
+import java.util.ArrayList;
+import java.util.Random;
 
-    private final PlayerSymbol playerSymbol;
-    private final Mediator mediator;
+public class UnbeatablePlayer extends Player {
+
+    private Time time;
 
     public UnbeatablePlayer(PlayerSymbol playerSymbol, Time time, Mediator mediator) {
-        super(time);
+        this.time = time;
         this.playerSymbol = playerSymbol;
         this.mediator = mediator;
     }
 
     @Override
     public void requestMove(Board board) {
-        super.waitOneSecond();
+        waitOneSecond();
         mediator.receiveMove(strategicMove(board));
-    }
-
-    @Override
-    public PlayerSymbol getSymbol() {
-        return playerSymbol;
     }
 
     private int strategicMove(Board board) {
         if (board.allAvailableMoves().size() > 12) {
-            return super.randomMove(board.allAvailableMoves());
+            return randomMove(board.allAvailableMoves());
         } else {
             return new MiniMax(playerSymbol).execute(board);
         }
     }
-    
+
+    private void waitOneSecond() {
+        time.pause(1000);
+    }
+
+    private int randomMove(ArrayList<Integer> moves) {
+        Random random = new Random();
+        int i = random.nextInt(moves.size());
+        return moves.get(i);
+    }
+
 }
